+++ packages/sdk/src/components/Sidebar/Sidebar.stories.tsx
@@ -223,7 +223,7 @@ export const CollectionsLeft: Story = {
               />
             </>
           }
-          bodyClassName="pr-2.5 pb-3"
+          bodyClassName="pb-3"
         >
           <SidebarSections
             sections={collectionsSections()}